PROX MAX LIN

+++ include/config.hpp
@@ -3,7 +3,7 @@
 
 static const int kBlockSizeCUDA = 256;
 
-typedef float real;
-//typedef double real;
+//typedef float real;
+typedef double real;
 
 #endif
